Refactor dispmanx code for threaded vsync
All Dispmanx specific rendering code is now shared between teletext
and cea608.

To use it, first call render_start with the width and height, and
with two functions: the init function will be called once before
the render loop starts up and should fill in the first field of
data. The draw function will be called after every vsync to get
next data. Both functions first argument is a pointer to a uint8
buffer. The draw function also takes a parameter indicating whether
to draw the odd or even field (although it is not possible to know
which one will actually be displayed next.)

The render_start function returns a void *. When you want to stop
rendering, call render_stop and pass the void * back in.